requirements.txt
Flask>=2.0.0
python-dotenv>=0.21.0
requests>=2.26.0
pyautogen
openai>=0.27.0
autogen-agentchat==0.5.6
autogen-core==0.5.6
autogen-ext==0.5.6
autogen-ext[openai]
autogen-ext[mcp]
playwright
mcp-server-fetch
google==3.0.0
google-auth==2.39.0
googleapis-common-protos==1.70.0
googlesearch-python==1.3.0